Drawing on multiple layers
The user can now select different layers to draw on to.
Layer index is what is used to refer to the active layer. It is probably
a good idea to use this throughout and use something like Layer
Picture::GetActiveLayer(){return layers[m_activeLayerIndex];} to get the
actual layer object itself.